Added Scripts

+++ auto_setup.sh
@@ -6,6 +6,8 @@ echo "I Will Install Your Programs For You"
 
 echo "______Installing Required Tools_______"
 
+cd ~
+
 sudo apt update
 yes | sudo apt upgrade
 
@@ -42,21 +44,35 @@ yes | sudo apt-get install cmake
 
 echo "______Pycharm Setup_______"
 
-cd ~/Downloads
-wget "https://download.jetbrains.com/python/pycharm-professional-2024.3.1.1.tar.gz"
-sudo tar xzf pycharm-*.tar.gz -C /opt/
-sudo rm pycharm-*.tar.gz
-echo $"\nexport PATH="/opt/pycharm-2024.3.1.1/bin:$PATH""
+cd ~
+DIRECTORY="/opt/pycharm-2024.3.1.1/"
+
+if [ -d "$DIRECTORY" ]; then
+    cd ~/Downloads
+    wget "https://download.jetbrains.com/python/pycharm-professional-2024.3.1.1.tar.gz"
+    sudo tar xzf pycharm-*.tar.gz -C /opt/
+    sudo rm pycharm-*.tar.gz
+    echo $"\nexport PATH="/opt/pycharm-2024.3.1.1/bin:$PATH""
+else
+    echo "Pycharm is already installed ${DIRECTORY}"
+fi
 
 
 
 echo "______Clion Setup_________"
 
+cd ~
+DIRECTORY="/opt/clion-2024.3.1.1/bin"
+
+if [ -d "$DIRECTORY" ]; then
 cd ~/Downloads
 wget "https://download.jetbrains.com/cpp/CLion-2024.3.1.1.tar.gz"
 sudo tar xvzf CLion-*.tar.gz -C /opt/
 sudo rm CLion-*.tar.gz
-echo $"\nexport PATH="/opt/pycharm-2024.3.1.1/bin:$PATH""
+echo $"\nexport PATH="/opt/clion-2024.3.1.1/bin:$PATH""
+else
+    echo "CLion is already installed ${DIRECTORY}"
+fi
